feat: better icons
